Experimental code that uses CSS grid layout instead of tables
Tables are slightly better because they can squeeze into smaller
available widths. It is also trivial to implement border-collapse in
tables, as they have native support for that.

Grids are slightly better because the browser can automatically allocate
into new rows, simplifying the generation loop. It also means it's easy
to change the amount of columns dynamically.

Neither solution is perfect.