Added an alternative ANSI palette that uses a compact layout and exactly 256 colors
In Gimp, there are two ways to reduce an image to just the colors from a
palette:

**Image → Mode → Indexed**

<https://docs.gimp.org/en/gimp-image-convert-indexed.html>

It shows a dialog that lets the user pick any of the available palettes,
but only those that have at most 256 colors. The results look quite
good, the colors are matched in a nice way.

**Colors → Map → Palette Map**

<https://docs.gimp.org/en/plug-in-palettemap.html>

It uses the currently selected palette, and applies the changes
immediately without any options. It calculates the luminosity of each
pixel and maps the black to the first palette entry and the white to the
last palette entry. It can be good for heat maps, depth maps, or other
kind of images that map each point to a single dimension (i.e.
`f(x,y)→z`), but it cannot be used for anything else.

Thus, unfortunately it is impossible to reduce an image to the colors of
a palette, if the palette has more than 256 colors. I mean, it should be
possible to write a plugin to do that, but I'm not aware of any. (Maybe
there is one already.)

That all means it was previously impossible to use Gimp to reduce an
image to the ANSI color palette, even though it has less than 256 unique
colors. Aiming to fix that use-case, this compact layout was born. The
first four grayscale colors had to be moved to the end of the palette,
because 256 isn't divisible by 6.

Thanks to @nervo for pointing it out.

Closes #14